requirements.txt
# Core dependencies
boto3>=1.34.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Analysis
scikit-learn>=1.3.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
